+++ src/vparsepkg/parser.nim
@@ -559,6 +559,7 @@ template parse_parameter_or_localparam_declaration_tree(p: var Parser, result: P
 proc parse_localparam_declaration(p: var Parser): PNode =
    result = new_node(p, NkLocalparamDecl)
    expect_token(p, result, TkLocalparam)
+   add_comment(p, result)
    get_token(p)
    parse_parameter_or_localparam_declaration_tree(p, result)
 
@@ -566,6 +567,7 @@ proc parse_localparam_declaration(p: var Parser): PNode =
 proc parse_parameter_declaration(p: var Parser): PNode =
    result = new_node(p, NkParameterDecl)
    expect_token(p, result, TkParameter)
+   add_comment(p, result)
    get_token(p)
    parse_parameter_or_localparam_declaration_tree(p, result)